+++ scilab/modules/hdf5/src/cpp/H5VlenData.cpp
@@ -46,7 +46,7 @@ H5Object & H5VlenData::getData(const unsigned int size, const unsigned int * ind
 
     for (unsigned int i = 0; i < size; i++)
     {
-        pos += cumprod[i] * index[i];
+        pos += (int)cumprod[i] * index[i];
     }
 
     if (pos >= totalSize)
